Casting number as a float (to avoid type conflict)

+++ 0x00-java-hello_world/PrintFloatVariable.java
@@ -1,6 +1,6 @@
 public class PrintFloatVariable {
     public static void main(String [] args) {
-        double number = Math.PI ;
+        float number = (float) Math.PI ;
         System.out.printf("PI Value: %.3f\n", number);
     }
 }
